refactor ref testcase
