+++ src/test/java/com/alibaba/json/bvt/ref/RefTest4.java
@@ -1,4 +1,4 @@
-package com.alibaba.json.bvt.serializer;
+package com.alibaba.json.bvt.ref;
 
 import java.math.BigDecimal;
