+++ src/test/java/com/alibaba/json/bvt/ref/RefTest6.java
@@ -1,4 +1,4 @@
-package com.alibaba.json.bvt.serializer;
+package com.alibaba.json.bvt.ref;
 
 import junit.framework.TestCase;
